lightweight-app/requirements.txt
flask==3.0.0
kafka-python-ng==2.2.2
gunicorn==21.2.0
